+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added `function` field to `modify_durability_change` power type. Can be `floor`, `round`, or `ceiling`. Defaults to `floor`.
 - Added `self_target_bientity_action`, `self_nearby_bientity_action`, and `target_nearby_bientity_action` to `damage_nearby_on_hit` power type. (All optional).
 - Added `attacker_target_bientity_action`, `attacker_nearby_bientity_action`, and `self_nearby_bientity_action` to `damage_nearby_when_hit` power type. (All optional).
+- Added `should_damage_condition` to `freeze` power type.
 
 ### Bi-entity Condition Types
 - Added `compare_dimensions` bi-entity condition type.